requirements.txt
anyio==3.3.4
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
dataclasses==0.6
deprecation==2.1.0
gotrue==0.2.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==2.10
iniconfig==1.1.1
mysql-connector-python==8.0.27
packaging==21.3
pluggy==1.0.0
postgrest-py==0.5.0
protobuf==3.19.1
py==1.11.0
pyparsing==3.0.6
pyTelegramBotAPI==4.2.2
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
toml==0.10.2
urllib3==1.26.7
websockets==9.1
